Move Python function in web.py
